Fix provenance issues with `SyntaxBuilder` (#44)
`vec.as_mut_ptr_range().end` was used to get a pointer to the start of the vec allocation. This goes through `Vec::deref` to a slice, which only has provenance of the initialized parts of the vec, so writing to the uninit part of the vec is not allowed.